requirements_fer_worked.txt
absl-py==2.1.0
astroid==3.0.1
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.4
decorator==4.4.2
deepface==0.0.89
dill==0.3.7
distlib==0.3.7
facenet-pytorch==2.5.3
fer==22.5.1
ffmpeg==1.4
filelock==3.12.4
fire==0.6.0
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.43.1
fsspec==2023.9.2
gast==0.5.4
gdown==5.1.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.1
guidata==3.1.0
guiqwt==4.4.4
gunicorn==21.2.0
h5py==3.10.0
identify==2.5.30
idna==3.4
imageio==2.31.6
imageio-ffmpeg==0.4.9
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.15.0
kiwisolver==1.4.5
lapx==0.5.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
mtcnn==0.1.1
namex==0.0.7
networkx==3.1
nodeenv==1.8.0
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.8.1.78
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
platformdirs==3.11.0
pre-commit==3.5.0
proglog==0.1.10
protobuf==4.25.3
psutil==5.9.6
psycopg2==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.17.2
pylint==3.0.1
pyparsing==3.1.1
pyqt-tools==1.0.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
PySocks==1.7.1
python-dateutil==2.8.2
python-version==0.0.2
PythonQwt==0.11.2
pytz==2023.3.post1
PyYAML==6.0.1
QtPy==2.4.1
requests==2.31.0
requests-oauthlib==2.0.0
retina-face==0.0.16
rich==13.7.1
rsa==4.9
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
soupsieve==2.5
supervision==0.15.0
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tf_keras==2.15.1
thop==0.1.1.post2209072238
tomli==2.0.1
tomlkit==0.12.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.199
urllib3==2.0.7
virtualenv==20.24.5
Werkzeug==3.0.2
wrapt==1.14.1
